# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/app/products/{slug}": {
      "get": {
        "operationId": "StorefrontController_getProduct",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get storefront product details"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/events": {
      "post": {
        "operationId": "StorefrontController_trackEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackStorefrontEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Track storefront analytics event"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/cart/items": {
      "post": {
        "operationId": "StorefrontController_addCartItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add product variant to cart"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/cart/{cartId}": {
      "get": {
        "operationId": "StorefrontController_getCart",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currencyCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get cart details by id"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/cart/{cartId}/items/{variantId}": {
      "put": {
        "operationId": "StorefrontController_updateCartItem",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update cart item quantity"
          }
        },
        "tags": [
          "storefront"
        ]
      },
      "delete": {
        "operationId": "StorefrontController_removeCartItem",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currencyCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove cart item"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/recovery/{token}": {
      "get": {
        "operationId": "StorefrontController_recoverAbandonedCart",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolve abandoned cart recovery link and redirect to checkout"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/recovery/{token}/open": {
      "get": {
        "operationId": "StorefrontController_trackAbandonedCartRecoveryOpen",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track abandoned cart recovery email open pixel"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
...
    "/app/checkout": {
      "post": {
        "operationId": "StorefrontController_checkout",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opaque client-generated key (16-200 characters) scoped to storefront.checkout",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout active cart and create order"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/checkout/quote": {
      "post": {
        "operationId": "StorefrontController_checkoutQuote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutQuoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quote checkout totals with optional loyalty redemption"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/checkout/summary": {
      "post": {
        "operationId": "StorefrontController_checkoutSummary",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutQuoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Summarize checkout totals with fulfillment selection"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/restock/track/{token}": {
      "get": {
        "operationId": "StorefrontController_trackRestock",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track restock notification click and redirect to product page"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
    "/app/orders/{orderCode}/track": {
      "get": {
        "operationId": "StorefrontController_trackOrder",
        "parameters": [
          {
            "name": "orderCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track order by order code"
          }
        },
        "tags": [
          "storefront"
        ]
      }
    },
...
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create category"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List categories"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
...
    "/filters": {
      "post": {
        "operationId": "FiltersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFilterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create filter"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "filters"
        ]
      },
      "get": {
        "operationId": "FiltersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List filters"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "filters"
        ]
      }
    },
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create product"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List products"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
...
    "/products/{productId}": {
      "get": {
        "operationId": "ProductsController_getById",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get product details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update product"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "General Ecommerce API",
    "description": "General Ecommerce backend APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TrackStorefrontEventDto": {
        "type": "object",
        "properties": {}
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {}
      },
      "CheckoutDto": {
        "type": "object",
        "properties": {}
      },
      "CheckoutQuoteDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateFilterDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}